requirements.txt
django
djangorestframework
langchain
langchain-google-genai
langchain-openai
langchain-groq
elevenlabs
python-dotenv
itsdangerous
mysqlclient
python-dotenv